+++ packages/components/src/components/Tag/Tag.vue
@@ -1,48 +1,48 @@
-<script>
+<script setup lang="ts">
 import CheckIcon from "@knime/styles/img/icons/check.svg";
 
-export default {
-  components: {
-    CheckIcon,
-  },
-  props: {
-    /**
-     * If the tags should have hover + cursor styles.
-     */
-    clickable: {
-      type: Boolean,
-      default: false,
-    },
-    /**
-     * If the tag should be displayed as active
-     */
-    active: {
-      type: Boolean,
-      default: false,
-    },
-  },
-};
+interface Props {
+  /**
+   * If the tags should have hover + cursor styles.
+   */
+  clickable?: boolean;
+  /**
+   * If the tag should be displayed as active
+   */
+  active?: boolean;
+}
+
+withDefaults(defineProps<Props>(), {
+  clickable: false,
+  active: false,
+});
 </script>
 
 <template>
-  <span :class="['tag', { clickable, active }]">
+  <span tabindex="0" :class="['tag', { clickable, active }]">
     <slot /><CheckIcon v-if="clickable && active" class="checkmark" />
   </span>
 </template>
 
 <style lang="postcss" scoped>
+@import url("@knime/styles/css/mixins.css");
+
 .tag {
   border: 1px solid var(--knime-dove-gray);
   margin-right: 5px;
   margin-bottom: 5px;
   padding: 3px 5px;
-  font-size: 14px;
-  line-height: 17px;
+  font-size: 13px;
+  line-height: 13px;
   display: inline-flex;
   justify-content: center;
   align-items: center;
   font-weight: 500;
   color: var(--knime-dove-gray);
+
+  &:focus-visible {
+    @mixin focus-outline 2, 3;
+  }
 
   & svg {
     position: relative;
@@ -58,8 +58,8 @@ export default {
 
     &:hover {
       color: var(--knime-white);
-      background-color: var(--knime-masala);
-      border-color: var(--knime-masala);
+      background-color: var(--knime-dove-gray);
+      border-color: var(--knime-dove-gray);
 
       & svg {
         stroke: var(--knime-white);