+++ packages/rte/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @knime/rich-text-editor
 
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [8d1c239]
+  - @knime/components@1.0.9
+
 ## 1.0.9
 
 ### Patch Changes